TileMapViewer: resize itself and destroy/reconstruct widgets on video mode change (a premiere). GUI: Fixed the code to allow resizing the game window (but no interface for it yet).

+++ branches/20110530-allegro5-sound/srcs/meka.h
@@ -280,7 +280,7 @@ struct t_meka_configuration
     bool    start_in_gui;
 
     // Applet: Game Screen
-    int     game_screen_scale;
+    float   game_screen_scale;
 
     // Applet: File Browser
     bool    fb_close_after_load;
